requirements.txt
aiohttp==3.9.2
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.2.0
appdirs==1.4.4
attrs==23.2.0
beautifulsoup4==4.12.2
bidict==0.22.1
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
convertdate==2.4.0
deprecation==2.1.0
fastapi==0.108.0
frozendict==2.4.0
frozenlist==1.4.1
gotrue==2.1.0
h11==0.14.0
html5lib==1.1
httpcore==0.17.3
httpx==0.24.1
idna==3.6
Jinja2==3.1.2
kombu==5.3.4
logsnag==1.0.1
lunardate==0.2.2
lxml==5.0.0
MarkupSafe==2.1.3
multidict==6.0.4
multitasking==0.0.11
numpy==1.26.3
packaging==23.2
pandas==2.1.4
peewee==3.17.0
postgrest==0.13.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.8.0
pyluach==2.2.0
PyMeeus==0.5.12
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==3.14.2
python-socketio==4.6.0
pytz==2023.3.post1
realtime==1.0.2
redis==5.0.1
requests==2.31.0
resend==0.6.0
sec-api==1.0.17
six==1.16.0
sniffio==1.3.0
socketIO-client==0.7.2
soupsieve==2.5
starlette==0.32.0.post1
storage3==0.7.0
StrEnum==0.4.15
stripe==7.10.0
supabase==2.3.4
supafunc==0.3.1
twilio==8.12.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
uvicorn==0.25.0
vine==5.1.0
wcwidth==0.2.12
webencodings==0.5.1
websocket-client==1.7.0
websockets==11.0.3
workalendar==17.0.0
yarl==1.9.4
yfinance==0.2.33
